requirements.txt
numpy==1.26.4
scikit-learn==1.5.2
statsmodels==0.14.4
scipy==1.12.0
fedot==0.7.4
pandas==2.2.3
PyYAML==6.0.2

# API
fastapi==0.115.9
python-multipart==0.0.20
uvicorn==0.34.0
psutil==7.0.0
pydantic-settings==2.10.1

# Config
hydra-core==1.3.2

# Database
alembic==1.16.1
SQLAlchemy==2.0.38
asyncpg==0.30.0
# aioodbc==0.5.0

# Tests
coverage==7.6.12
